Fix an issue reading transaction logs over network.
NetworkReadableLogChannel should not check if there is enough data available before reading.
Netty will throw an IndexOutOfBoundsException if the network stream doesn't supply enough data
to match the requested type.

ReadPastEndException was refactored as a singleton without stack trace as it is a common and
expected case in the code paths.

Also removed some logging previously added for debugging this issue.