Use same property equality logic in tx-state.
When looking up indexed nodes from the transaction state, we need to do
the same type coercions of property values that we do when looking up
nodes from the committed indexes. This is done by using the equality
defined in DefinedProperty.

Along with this the coercions made in DefinedProperty have been made
more explicit, and comparisons between integral numbers and floating
point numbers have been made more strict, so that only exact comparison
is ever made.