Reduce probability of false failure of pageFaultForRead* test
The problem is that it can race with the optimistic "self-healing" of the cache
and therefor have a highly variable run time.

It would be better to stabilise the running time, but my attempts at doing that
have thus far only led to subtly breaking other things.
So this needs further investigation. In the mean time, this will break fewer
pull request builds.