Exceptions in eviction thread now wakes up blocked page faulters.
The problem was that a page faulting thread could make it past the
`assertHealthy()` check, and then park, and then the eviction thread could
encounter an exception and then never tell the parked threads about it. They
would be stuck for however long it woudl take for the eviction thread to
finally evict and free a page for them.

Now, if the eviction thread knows that threads are parked, and it failes to
evict a page, it will instead signal to the parked thread and transfer the
exception to it, instead of a free page. The parked thread will then wake up
and throw the exception.