Fix failure of backing up store upgraded from old backup.
Trying to get more than one transaction during backup from
an upgraded store might fail since they might contain only
a single transaction in its backed up log. This is the case
for 1.9 stores currently.

The commit basically reverts this mismatch and only tries to
backup a single transaction. A more robust and flexible solution
could look at the logs and backup a reasonable amount of them
if available, which might be good for debugability.