Fixes bug around flattening of plan descriptions
InternalPlanDescription toSeq method was returning a flattened list of all plans, but doing so by
relying on children.toSeq to flatten all children. This was incorrect behavior for Children.toSeq, which
was corrected to only return a Seq of the children, and then the InternalPlanDescription.toSeq was
replaced with a correctly behaving flatten method.