Abbreviate common connection closed error.
When the client closes a connection to the server before recieving the full
reply, the server currently prints a lengthy stack trace to the log. This
stack trace is not helpful, because we already do the right thing here.

Replace stack trace with one-line log message noting the client closed the
connection.