Fix flakiness: UniqueConstraintStressIT
The expected result set was not sufficiently large; specifically
creating the constraint can fail and cause some of the node creation
to fail as well, not due to a constraint violation, but due to a
concurrent schema creation on master. The old code had several
logical faults.

The slave transactions can also fail due to a ComException, due to the
RequestContext used for the lock-session on master already being
in use. This is a known and expected behaviour in itself, but it
is not really known why it occurs in this test since it only has one
thread doing slave transactions. The strategy so far is to catch
and ignore it, but it is probably worth investigating and explaining.