Remove the use of reachability fences
We don't need them because the use of the native pointer is contained within the critical section of the pin, as represented by the StampedLock. This locking similarly constrains the life-time of the page objects.
In essence, use of the native pointer cannot move above taking the lock, nor bellow its release, and since the MuninnPage objects are themselves StampedLocks, their life-times is guaranteed to extend at least as far as their last lock release.