Decouple compiler version and planner version
To control language version `CYPHER x.y` and to control
planner we have
```
PLANNER COST
PLANNER RULE
```
By default we use a variant that uses what we currently is known to be safe in COST and
fallback to RULE when we are at risk of creating slow plans. Currently we don't do varlength,
cycles and updates in COST.

The default planner can also be configured by `query.planner.version`.