Immutable information in RelationshipProxy
Since the start node, end node, and type of a relationship never
changes, this information can be kept within the RelationshipProxy
object. Doing so does not require significant storage overhead, with the
current address sizes (nodeId=35bit, relationshipId=35bit, typeId=16bit)
this information (together with the id of the relationship) occupies 212
bits, which can be stored in 16 bytes (128 bits). The implementation in
this changeset packs the high bits of the start/end node ids and the
relationship id into separate nibbles of a 16 bit (short) fields.

This limits the number of roundtrips required down to the store layer.