Fixed a bug where last committed tx id in backup store logs are wrongly set
, which results in slaves failing to start in ha from backup db.

The bug arises when performing online backup and data writing concurrently.
As there is a time span between the start and the end of the store copy,
if more tx are flushed into stores during this time span, the last committed
tx id recorded in stores becomes different from the tx id before store copy.
Finally we get logs with wrong headers.