Simplify transaction log force batching
Instead of having a dedicated fsync batching thread, we now use a flat-combining inspired approach where force requests pile up, and whoever happens to grab the lock, gets to do the forcing for the whole pile at time.

This dramatically reduces the number of system calls that are involved in contended transaction log forcing, and makes the code easier to profile and also slightly faster.