Considers all deleted relationships when patching chain positions
Previously when multiple consecutive relationships in a chain got deleted
the current chain positions for cached nodes (NodeImpl) got patched to go
the next relationship in that chain. Problem was that the chain
positions got patched by looking at each deleted relationship individually
and so depending on the order the relationships existed in the chain, the
chain positions might have gotten patched to another deleted relationship.

This commit collects deleted relationships and patches each chain position
knowing all deleted relationships and so these "holes" can be detected and
the next relationship after the hole selected.