Fixes memory leak around page swapping
observed when doing frequent flushing of the page cache. The problem was
that the translation table of a paged file was still alive due to being
referenced by a page swapper, via an eviction callback.

Setting fields to null on the page swapper on close makes these
translation tables collectable.